// src/app/private/menu/models/categoria-response.model.ts
export class CategoriaResponse {
  id: number = 0;
  cantidadDeProductos: number = 0;
  precioMaximo: number = 0;
  precioMinimo: number = 0;
  nombre: string = "";
  descripcion: string = "";
  disponibilidad: boolean = true;
  disponibilidadDescripcion: string = "";
  estado: boolean = true;
  estadoDescripcion: string = "";
  urlImagen: string = "";
}